+++ vector/src/test/java/im/vector/app/test/fakes/FakeSharedSecretStorageService.kt
@@ -75,7 +75,12 @@ class FakeSharedSecretStorageService : SharedSecretStorageService by mockk() {
 
     override fun checkShouldBeAbleToAccessSecrets(secretNames: List<String>, keyId: String?) = integrityResult
 
+    @Deprecated("Requesting custom secrets not yet support by rust stack, prefer requestMissingSecrets")
     override suspend fun requestSecret(name: String, myOtherDeviceId: String) {
+        TODO("Not yet implemented")
+    }
+
+    override suspend fun requestMissingSecrets() {
         TODO("Not yet implemented")
     }